# Conflicts:
#	src/main/resources/data.sql

+++ src/main/java/com/example/cms/controller/exceptions/RoomNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.cms.controller.exceptions;
 
 public class RoomNotFoundException extends RuntimeException{
-    public RoomNotFoundException(Long id) {
+    public RoomNotFoundException(Integer id) {
         super("Could not find room " + id);
     }
 }
